+++ R/geom_subgene_label.R
@@ -114,6 +114,12 @@ GeomSubgeneLabel <- ggplot2::ggproto(
     subgroup = NA
   ) {
 
+    # Detect flipped coordinates
+    coord_flip <- inherits(coord, "CoordFlip")
+    if (coord_flip) {
+      data$angle <- data$angle + 90
+    }
+
     # Convert aesthetic names
     data$xmin <- data$xsubmin
     data$xmax <- data$xsubmax
@@ -129,18 +135,33 @@ GeomSubgeneLabel <- ggplot2::ggproto(
     }
 
     # Use ggfittext's fittexttree to draw the text
-    gt <- grid::gTree(
-      data = data,
-      padding.x = padding.x,
-      padding.y = padding.y,
-      place = align,
-      min.size = min.size,
-      grow = grow,
-      reflow = reflow,
-      cl = "fittexttree",
-      height = height,
-      fullheight = TRUE
-    )
+    if (coord_flip) {
+      gt <- grid::gTree(
+        data = data,
+        padding.x = padding.x,
+        padding.y = padding.y,
+        place = align,
+        min.size = min.size,
+        grow = grow,
+        reflow = reflow,
+        cl = "fittexttree",
+        width = height,
+        fullheight = TRUE
+      )
+    } else {
+      gt <- grid::gTree(
+        data = data,
+        padding.x = padding.x,
+        padding.y = padding.y,
+        place = align,
+        min.size = min.size,
+        grow = grow,
+        reflow = reflow,
+        cl = "fittexttree",
+        height = height,
+        fullheight = TRUE
+      )
+    }
     gt$name <- grid::grobName(gt, "geom_subgene_label")
     gt
   }